+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallEvent.swift
@@ -1102,4 +1102,43 @@ enum InternalSuperwallEvent {
       ]
     }
   }
+
+  enum PermissionState {
+    case requested
+    case granted
+    case denied
+  }
+
+  struct Permission: TrackableSuperwallEvent {
+    let state: PermissionState
+    let permissionName: String
+    let paywallIdentifier: String
+    var superwallEvent: SuperwallEvent {
+      switch state {
+      case .requested:
+        return .permissionRequested(
+          permissionName: permissionName,
+          paywallIdentifier: paywallIdentifier
+        )
+      case .granted:
+        return .permissionGranted(
+          permissionName: permissionName,
+          paywallIdentifier: paywallIdentifier
+        )
+      case .denied:
+        return .permissionDenied(
+          permissionName: permissionName,
+          paywallIdentifier: paywallIdentifier
+        )
+      }
+    }
+    var audienceFilterParams: [String: Any] = [:]
+
+    func getSuperwallParameters() async -> [String: Any] {
+      return [
+        "permission_name": permissionName,
+        "paywall_identifier": paywallIdentifier
+      ]
+    }
+  }
 }