+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 ### Enhancements
 
 - Adds `paywallPreload_start` and `paywallPreload_complete` events.
+- Adds `request permission` action support allowing you to request Notification permissions from paywalls.
 
 ## 4.11.2